+++ wrangler.toml
@@ -11,15 +11,30 @@ command = "npm run build"
 account_id = "fffa4b4363a7e5250af8357087263b3a" # DAG House CF account
 #route = "claudio.web3.storage"
 
+[env.production.durable_objects]
+bindings = [
+  { name = "WEBSOCKET_SESSIONS", class_name = "WebSocketSession0" }
+]
+
 # Staging!
 [env.staging]
 # name = "minibus-staging"
 account_id = "fffa4b4363a7e5250af8357087263b3a" # DAG House CF account
 #route = "claudio-staging.web3.storage"
 
+[env.staging.durable_objects]
+bindings = [
+  { name = "WEBSOCKET_SESSIONS", class_name = "WebSocketSession0" }
+]
+
 # Test!
 [env.test]
 workers_dev = true
+
+[env.test.durable_objects]
+bindings = [
+  { name = "WEBSOCKET_SESSIONS", class_name = "WebSocketSession0" }
+]
 
 [env.alanshaw]
 workers_dev = true
@@ -28,9 +43,23 @@ account_id = "4fe12d085474d33bdcfd8e9bed4d8f95"
 [env.alanshaw.vars]
 DEBUG = "true"
 
+[env.alanshaw.durable_objects]
+bindings = [
+  { name = "WEBSOCKET_SESSIONS", class_name = "WebSocketSession0" }
+]
+
 [env.vsantos]
 workers_dev = true
 account_id = "7ec0b7cf2ec201b2580374e53ba5f37b"
 
 [env.vsantos.vars]
 DEBUG = "true"
+
+[env.vsantos.durable_objects]
+bindings = [
+  { name = "WEBSOCKET_SESSIONS", class_name = "WebSocketSession0" }
+]
+
+[[migrations]]
+tag = "v0"
+new_classes = ["WebSocketSession0"]
